get the current date by index rather than parsing the item text
Get the current date by index to improve reliability of
dates based on the calendar of the current year. This fixes
a problem where dates would not be selected correctly when
the month of the current year has a different number of days
from the year 1970, for example in a leap year.

fixes #254